Update resolve_address.h
The use of size_t for the length creates an issue for a big Endian architecture that defines size_t as a 64 bit variable.  The function grpc_tcp_server_prepare_socket calls getsockname casting the length of the addr structure in len as (socklen_t*)&sockname_temp.len so the top half of the 64 bit len gets passed to getsockname which is 0.  The call to getsockname completes successfully and an invalid sockname_temp is passed to grpc_sockaddr_get_port contains an invalid sa_family.